ml_training/requirements.txt
tensorflow
numpy<2.0.0
pillow
tensorflow-datasets
matplotlib
